% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_n_proc_admisssion.R
\name{add_n_proc_admission}
\alias{add_n_proc_admission}
\title{add_n_proc_admission}
\usage{
add_n_proc_admission(data)
}
\arguments{
\item{data}{medicare data}
}
\value{
medicare analytic file with one additional variable "n_cpt_admission".
}
\description{
Add new variable "n_cpt_admission" to indicate number of CPTs in one hospital admission
using member_id, facility claim admission and discharge dates
}
