% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_all_filters.R
\name{apply_all_filters}
\alias{apply_all_filters}
\title{Use all cohort filters}
\usage{
apply_all_filters(data, remove_multi_procedure = T, ...)
}
\arguments{
\item{data}{medicare}

\item{remove_multi_procedure}{TRUE: remove all mult procedure cases;
FALSE: add variable "n_cpt_admission"}

\item{...}{parameters inputs for all sub-filter functions:
remove_multi_proc, addd_n_proc_admission, remove_nonUS_trained, remove_fellowship, remove_non_complete}
}
\value{

}
\description{
Use all cohort filters
}
\details{
filters including: remove multi procedures,
remove non-us med school graduates,
remove fellowship trained surgeons,
remove non-complete cases
}
