% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_selection_diagram.R
\name{cohort_selection_diagram}
\alias{cohort_selection_diagram}
\title{Cohort selection diagram for defined filters}
\usage{
cohort_selection_diagram(data, add_variables = model_vars, ...)
}
\arguments{
\item{data}{top data (raw data)}

\item{add_variables}{variables that need to be used to exclude missing values, for example
years of experience. If need to add multiple variables, use syntax c("var1", "var2")}

\item{...}{}
}
\value{

}
\description{
This is often used to understand how many cases and surgeons are removed in each filter
}
