% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_formula.R
\name{create_model_formulas}
\alias{create_model_formulas}
\title{Create model formulas}
\usage{
create_model_formulas(
  targets = outcomes,
  primary_covariate = primary,
  other_covariates = covariates,
  random_effects = "id_surgeon",
  interaction_term = NULL,
  mgcv = FALSE
)
}
\arguments{
\item{targets}{character vector of target variable names}

\item{primary_covariate}{character string of the key predictor variable}

\item{other_covariates}{character vector of other predictors}

\item{random_effects}{names of the random effect variables. Default is id_surgeon.}

\item{interaction_term}{a character vector}

\item{mgcv}{logical. Convert random effects specification to mgcv style? Default is false.}
}
\value{
a list of model formulas
}
\description{
Create model formulas
}
