% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_procedures_table.R
\name{multiple_procedures_table}
\alias{multiple_procedures_table}
\title{multiple procedures in one admission frequency table}
\usage{
multiple_procedures_table(
  data,
  project_sepcific_procedure_tax = score_operation_procedure
)
}
\arguments{
\item{data}{}

\item{project_sepcific_procedure_tax}{project specific procedure taxinomy}
}
\value{
frequency table
}
\description{
make frequency table for multiple procedures combinations in one admission.
By making the frequency table, it's easier to see what procedures happened in on admission
more frequently
}
