% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_selection_diagram.R
\name{n_case_surg}
\alias{n_case_surg}
\title{Number of cases and surgeons in the dataset}
\usage{
n_case_surg(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Number of cases and surgeons in the dataset
}
\details{
surgeon variable name shoule be id_physician_npi
}
