% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data_for_model.R
\name{prep_data_for_model}
\alias{prep_data_for_model}
\title{Prepare data for analysis with a statistical model}
\usage{
prep_data_for_model(
  data,
  standardize = c("age_at_admit", "AHRQ_score", "val_hosp_rn2bed_ratio",
    "val_hosp_mcday2inptday_ratio", "val_hosp_rn2inptday_ratio")
)
}
\arguments{
\item{data}{The current data.}

\item{standardize}{Character string of variables to standardize (mean 0,
stddev 1)}
}
\value{
A data frame worthy of applying a statistical model to.
}
\description{
Prepare data for analysis with a statistical model
}
\details{
create binary variables for models, including ses, race, admission status.
Standardized continuous variables for easy converging; rename variables that need to
be used in models.
}
