% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_n_proc_admisssion.R
\name{remove_multi_ecv_proc}
\alias{remove_multi_ecv_proc}
\title{Remove multiple ECV procedures and keep one unique ECV case}
\usage{
remove_multi_ecv_proc(data)
}
\arguments{
\item{data}{medicare data}
}
\description{
Remove multiple ECV procedures and keep one unique ECV case
}
