% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_n_proc_admisssion.R
\name{remove_multi_proc}
\alias{remove_multi_proc}
\title{remove multiple procedures in one admission cases}
\usage{
remove_multi_proc(data)
}
\arguments{
\item{data}{medicare data}
}
\description{
remove multiple procedures in one admission cases
}
