% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_nonUS_trained.R
\name{remove_nonUS_trained}
\alias{remove_nonUS_trained}
\title{Remove Non-US Trained surgeons}
\usage{
remove_nonUS_trained(data)
}
\arguments{
\item{data}{medicare data}
}
\value{
medicare data without non-US trained surgeons
}
\description{
use medciare variable us_medschool based on ABS and AMA
}
