% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_frac_group.R
\name{sample_frac_group}
\alias{sample_frac_group}
\title{Sample by a proportion of a grouping variable}
\usage{
sample_frac_group(data, proportion = 0.1, id = id_physician_npi, seed = 1234)
}
\arguments{
\item{data}{A data frame.}

\item{proportion}{A value between 0 and 1.}

\item{id}{The bare column name containing the grouping variable whose labels you wish to sample.}
}
\value{
A filtered data frame containing roughly the proportion of id labels provided.
}
\description{
Sample by a proportion of a grouping variable
}
\details{
when the data is too big, we sample the data (usually
}
