% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_all_filters.R
\name{apply_all_filters}
\alias{apply_all_filters}
\title{Use all cohort filters}
\usage{
apply_all_filters(
  data,
  remove_multi_procedure = T,
  fellowship_council_data_path,
  medicare_gs_list_path,
  variables,
  yoe_cutoff = 35,
  ...
)
}
\arguments{
\item{data}{medicare}

\item{remove_multi_procedure}{TRUE: remove all mult procedure cases;
FALSE: add variable "n_cpt_admission"}

\item{fellowship_council_data_path}{path to fellowship council dt}

\item{medicare_gs_list_path}{assign value to NULL if you don't want use the
medicare specialty code criteria.}

\item{variables}{variables that need to be used to exclude missing values, for example
years of experience. If need to add multiple variables, use syntax c("var1", "var2")}

\item{yoe_cutoff}{the max years of experience included in the dataset}

\item{...}{parameters inputs for all sub-filter functions:
remove_multi_proc, addd_n_proc_admission, remove_nonUS_trained, remove_fellowship, remove_non_complete}
}
\value{

}
\description{
Use all cohort filters
}
\details{
filters including: remove multi procedures,
remove non-us med school graduates,
remove fellowship trained surgeons,
remove non-complete cases
}
