% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_selection_diagram.R
\name{cohort_selection_diagram}
\alias{cohort_selection_diagram}
\title{Cohort Selection Diagram}
\usage{
cohort_selection_diagram(data, variables = model_vars, yoe_cutoff = 35, ...)
}
\arguments{
\item{data}{original data}

\item{variables}{variables that need to be used to exclude missing values, for example
years of experience. If need to add multiple variables, use syntax c("var1", "var2")}

\item{yoe_cutoff}{the max years of experience included in the dataset}

\item{...}{}
}
\value{

}
\description{
create cohort selection diagram based on apply_all_filters function.
This is often used to understand how many cases and surgeons are removed in each filter
}
